Try blueocean implementation classes moved to blueocean-rest-impl
- Removed try.js as it's not more needed
