+++ blueocean-rest-impl/src/main/webapp/images/24x24/blueocean.png
