+++ blueocean-rest-impl/src/main/webapp/images/48x48/blueocean.png
