+++ blueocean-web/gulpfile.js
@@ -60,17 +60,6 @@ builder.bundle('src/main/js/blueocean.js')
     .generateNoImportsBundle();
 
 //
-// Create the "Try Blue Ocean" Javascript bundle.
-// This .js bundle will be added to every classic Jenkins page
-// via a PageDecorator. Using this as a way of enticing Jenkins
-// users to move from classic Jenkins to Blue Ocean where possible.
-//
-builder.bundle('src/main/js/try.js')
-    .inDir('target/classes/io/jenkins/blueocean')
-    .import('jquery-detached', 'core-assets/jquery-detached:jquery2') // Bundled in Jenkins 2.x
-    .less('src/main/less/try.less');
-
-//
 // An Internet Explorer specific polyfill.
 // Go IE ... you never fail to make me smile :(
 //
